+++ library/src/main/java/au/edu/griffith/caffeinatedmap/utils/Distance.java
@@ -2,14 +2,14 @@ package au.edu.griffith.caffeinatedmap.utils;
 
 public class Distance {
 
-    public static double getHypotenuse(double x1, double y1, double x2, double y2) {
-        double x = getAbsolute(x1 - x2);
-        double y = getAbsolute(y1 - y2);
+    public static double getHypotenuse(int x1, int y1, int x2, int y2) {
+        int x = getAbsolute(x1 - x2);
+        int y = getAbsolute(y1 - y2);
         return Math.sqrt(x * x + y * y);
     }
 
-    private static double getAbsolute(double value) {
-        return (value <= 0.0) ? 0.0 - value : value;
+    private static int getAbsolute(int value) {
+        return (value <= 0) ? 0 - value : value;
     }
 
 }
